change window resolution to 1280*800

+++ src/gui/Point.java
@@ -1,31 +1,31 @@
 package gui;
 
 public class Point {
-	public int x;
-	public int y;
+	public double x;
+	public double y;
 	public int row;
 	
 	public Point() {
 		x = 0; y = 0; row = 0;
 	}
 	
-	public Point(int x, int y) {
+	public Point(double x, double y) {
 		this.x = x;
 		this.y = y;
 		this.row = 0;
 	}
 	
-	public Point(int x, int y, int row) {
+	public Point(double x, double y, int row) {
 		this.x = x;
 		this.y = y;
 		this.row = row;
 	}
 	
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 	
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
